Implemented Pundit and basic roles

+++ app/controllers/posts_controller.rb
@@ -10,13 +10,13 @@ class PostsController < ApplicationController
   def new
     @topic = Topic.find(params[:topic_id])
     @post = Post.new
-    authorize! :create, Post, message: "You need to be a member to create a new post."
+    authorize @post
   end
 
   def edit
     @topic = Topic.find(params[:topic_id])
     @post = Post.find(params[:id])
-    authorize! :edit, @post, message: "You need to own the post to edit it."
+    authorize @post
   end
 
   def create
@@ -24,8 +24,7 @@ class PostsController < ApplicationController
     @post = current_user.posts.build(post_params)
     @post.topic = @topic
 
-    authorize! :create, @post, message: "You need to be signed up to do that."
-
+    authorize @post
     if @post.save
       redirect_to [@topic, @post], notice: "Post was saved successfully."
     else
@@ -38,7 +37,8 @@ class PostsController < ApplicationController
   def update
     @topic = Topic.find(params[:topic_id])
     @post = Post.find(params[:id])
-    authorize! :update, @post, message: "You need to own the post to edit it."
+    authorize @post
+
     if @post.update(post_params)
       redirect_to [@topic, @post], notice: "Post was saved successfully."
     else
@@ -50,9 +50,9 @@ class PostsController < ApplicationController
   def destroy
     @topic = Topic.find(params[:topic_id])
     @post = Post.find(params[:id])
-
     title = @post.title
-    authorize! :destroy, @post, message: "You need to own the post to delete it."
+
+    authorize @post
     if @post.destroy
       flash[:notice] = "\"#{title}\" was deleted successfully."
       redirect_to @topic
